adding math

+++ malloc/malloc.h
@@ -1,6 +1,8 @@
 #ifndef _HALLOCATOR
 #define _HALLOCATOR
 
+#include "../stddef/stddef.h" // size_t
+
 /*
  * Memory management library
  * Something as in stdlib
@@ -17,14 +19,14 @@
 
 typedef struct __node_t
 {
-    int              size;
+    size_t           size;
     struct __node_t *next;
 } node_t;
 
 typedef struct
 {
-    int size;
-    int magic;
+    size_t size;
+    int    magic;
 } header_t;
 
 /*
@@ -32,8 +34,7 @@ typedef struct
  * Alloc 'size' bytes and return a pointer to the first byte
  * If cant alloc, return NULL
  */
-void *
-malloc(int size);
+void *malloc(size_t size);
 
 /*
  * void *rehalloc(void* ptr, int size)
@@ -41,28 +42,24 @@ malloc(int size);
  * returns a pointer to a new possition and free previous pointer.
  * On error previous pointer is not free and return NULL
  */
-void *
-realloc(void *ptr, int size);
+void *realloc(void *ptr, size_t size);
 
 /*
  * void fhree(void* ptr)
  * Free memory pointed by ptr
  */
-void
-free(void *ptr);
+void free(void *ptr);
 
 /*
  * Automatically initialized before main(), can also be
  * called, it checks if it was initialized before
  */
-void
-__halloc_init();
+void __halloc_init();
 
 /*
  * Automatically call at exit, can also be called manually
  * but calls to allocator after this function would end in a segfault
  */
-void
-__alloc_destroy();
+void __alloc_destroy();
 
 #endif // !_HALLOCATOR
